requirements.txt
tensorflow==2.20.0
streamlit==1.38.0
numpy
pillow
scipy